Option: Collect metrics for all reconciled objects
At this point the controller runtime collects metrics for all objects it
reconciles. While useful, this doesn't allow one to investigate if there
are infinitely requeued objects, objects that are slow to reconcile,
ones that never reconciled etc.

Since collecting metrics at this granularity leads to high cardinality,
I'm proposing for this to be optional (and thus trivially backwards compatible.

This example only covers one of the metrics, but all the other ones
would be treated equally.

There are some edge cases to cover:
- what about metrics that are incremented in a `defer` statement before
  we find out what the object name is (since it could be invalid)
- we init some metrics by adding zero - this could be done for the `all`
  catchall label, but do we want to have this one present in the first
place? If we're collecting metrics per object, we should discard this
series in favour of summing everything else
- what about gc? How do we treat reconciled objects that got deleted?